Build images using UV
For backward compatibility we build two sets of images

1) We don't change the way how we build images using PIP.
   This assumes having so called requirements image
   that contains all the necessary wheels and then we
   build all other project images using these wheels
   from the requirements image.

2) Also we add another set of images that we build
   using the two stage Dockerfile. On the first stage
   we build venv using UV and then we just copy the
   venv as is to the project image.

Depends-On: Id6412055f742990f08aa1befa7a41f1670414d4c

+++ scripts/project_specific/horizon/01_configure_horizon.sh
@@ -8,7 +8,22 @@ DASHBOARD_ROOT=/usr/local/share/openstack_dashboard
 MANAGE_CMD="${SITE_PACKAGES_ROOT}/openstack_dashboard/manage.py"
 LOCAL_SETTINGS="${SITE_PACKAGES_ROOT}/openstack_dashboard/local/local_settings.py"
 
-cp /tmp/${PROJECT}/manage.py ${SITE_PACKAGES_ROOT}/openstack_dashboard/
+if [[ -f /tmp/${PROJECT}/manage.py ]]; then
+    cp /tmp/${PROJECT}/manage.py ${SITE_PACKAGES_ROOT}/openstack_dashboard/
+else
+    cat <<EOF > ${SITE_PACKAGES_ROOT}/openstack_dashboard/manage.py
+#!/usr/bin/env python
+import os
+import sys
+from django.core.management import execute_from_command_line
+
+if __name__ == "__main__":
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE",
+                        "openstack_dashboard.settings")
+    execute_from_command_line(sys.argv)
+EOF
+    chmod +x ${SITE_PACKAGES_ROOT}/openstack_dashboard/manage.py
+fi
 
 # wsgi/horizon-http needs open files here, including secret_key_store
 chown -R horizon ${SITE_PACKAGES_ROOT}/openstack_dashboard/local/
